requirements.txt
streamlit
easyocr
pillow
numpy
reportlab
torch
torchvision
